# gh-106524: Zero template items before initializing them in `_sre.template`