Fixes #106524

* made `_sre_template_impl` set items of 'self->items' array to zero before initializing them with group indices and literals
* added test


<!-- gh-issue-number: gh-106524 -->
* Issue: gh-106524
<!-- /gh-issue-number -->