@serhiy-storchaka 2023-07-07
Good catch. Other approach is to call `Py_SET_SIZE(self, i)` on an error during initialization. It may have less impact on performance. But how large this impact in the first place? If it is insignificant, `memset()` will do.