@chgnrdv 2023-07-07
@serhiy-storchaka, benchmarks speak for themselves:

`./python -m pyperf timeit -s 'import _sre' '_sre.template("", ["", 1, "", 1, "", 1, "", 1, ""])'`

main:
`Mean +- std dev: 480 ns +- 2 ns`

with memset:
`Mean +- std dev: 523 ns +- 7 ns`

with Py_SET_SIZE:
`Mean +- std dev: 474 ns +- 2 ns`