@miss-islington 2023-07-08
Sorry @chgnrdv and @serhiy-storchaka, I had trouble checking out the `3.11` backport branch.
                                Please retry by removing and re-adding the "needs backport to 3.11" label.
                                Alternatively, you can backport using [cherry_picker](https://pypi.org/project/cherry-picker/) on the command line.
                                ```
                                cherry_picker 2ef1dc37f02b08536b677dd23ec51541a60effd7 3.11
                                ```